#10 'Integer' in repositories was refactored to 'Long'

+++ src/main/java/kg/attractor/xfood/repository/CheckListRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface CheckListRepository extends JpaRepository<CheckList, Integer> {
+public interface CheckListRepository extends JpaRepository<CheckList, Long> {
 }
