+++ src/main/java/kg/attractor/xfood/repository/CriteriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CriteriaRepository extends JpaRepository<Criteria, Integer> {
+public interface CriteriaRepository extends JpaRepository<Criteria, Long> {
 }
